fix middleware and cleanup redundant function
