+++ python/packages/core/tests/openai/test_openai_responses_client.py
@@ -2,7 +2,6 @@
 
 import asyncio
 import base64
-import json
 import os
 from typing import Annotated
 from unittest.mock import MagicMock, patch
@@ -136,18 +135,6 @@ def test_init_with_default_header(openai_unit_test_env: dict[str, str]) -> None:
     for key, value in default_headers.items():
         assert key in openai_responses_client.client.default_headers
         assert openai_responses_client.client.default_headers[key] == value
-
-
-def test_openai_responses_client_instructions_sent_once(openai_unit_test_env: dict[str, str]) -> None:
-    """Ensure instructions are only included once for OpenAI Responses requests."""
-    client = OpenAIResponsesClient()
-    instructions = "You are a helpful assistant."
-    chat_options = ChatOptions(instructions=instructions)
-
-    prepared_messages = client.prepare_messages([ChatMessage(role="user", text="Hello")], chat_options)
-    request_options = client._prepare_options(prepared_messages, chat_options)  # type: ignore[reportPrivateUsage]
-
-    assert json.dumps(request_options).count(instructions) == 1
 
 
 @pytest.mark.parametrize("exclude_list", [["OPENAI_RESPONSES_MODEL_ID"]], indirect=True)
